Add files via upload
The Lagrange interpolation with different numbers of points was calculated for the sine and square functions. The Runge phenomenon was demonstrated while plotting the Lagrange interpolation for the function f(x) = \frac{1}{1+25x^2}. Then, Hermite interpolation was performed for the same function, which resulted in a lack of oscillations and a better approximation.